streamlit_app/requirements.txt
# Streamlit 앱 의존성

# Web Framework
streamlit==1.29.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
joblib==1.3.2

# Model Interpretation (optional)
# shap==0.43.0

# Excel/Report Generation (optional)
# openpyxl==3.1.2
# reportlab==4.0.7
